+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ## [1.1.6.0](https://github.com/unzerdev/php-sdk/compare/1.1.5.0..1.1.6.0)
 ### Added
 * Add payment type Paylater Invoice.
+* Add properties `companyType` and `owner` to `CompanyInfo` class.
+* Add shipping type to `Address` class.
 * Allow setting the clientIp manually.
 
 ### Changed